ui: consistent layout for summary contents
