+++ src/components/TaxDetails/TaxSummaryCard/taxSummaryCard.scss
@@ -1,9 +1,6 @@
 .Layer__TaxSummaryCard {
-  margin-block-end: var(--spacing-md);
-
-  &__Content {
-    width: 100%;
-  }
+  width: 100%;
+  border-block-end: 1px solid var(--color-base-200);
 
   &__Overview {
     min-width: 200px;
@@ -14,6 +11,7 @@
   &__Breakdown {
     flex: 1;
     padding: var(--spacing-lg);
+    padding-block-end: var(--spacing-xl);
   }
 
   &__Section {
@@ -24,6 +22,52 @@
       padding-block-end: 0;
       border-block-end: none;
     }
+  }
+
+  &__Grid {
+    display: grid;
+    grid-template-columns: auto auto auto auto auto;
+    row-gap: var(--spacing-3xs);
+    column-gap: var(--spacing-md);
+    align-items: baseline;
+    width: fit-content;
+
+    > .Layer__Span {
+      text-align: right;
+    }
+
+    &--mobile {
+      grid-template-columns: 1fr auto 1fr auto 1fr;
+      column-gap: var(--spacing-sm);
+      width: 100%;
+    }
+  }
+
+  &__SectionGroup {
+    display: contents;
+
+    > .Layer__Span:first-child {
+      grid-column: 1 / -1;
+      text-align: left;
+    }
+
+    &:not(:first-child) > .Layer__Span:first-child {
+      padding-block-start: var(--spacing-md);
+      margin-block-start: var(--spacing-3xs);
+      border-block-start: 1px solid var(--color-base-100);
+    }
+  }
+
+  &--mobile {
+    width: 100%;
+  }
+
+  &__OverviewCard {
+    padding: var(--spacing-lg);
+  }
+
+  &__BreakdownCard {
+    padding: var(--spacing-lg);
   }
 }
