+++ src/components/TaxDetails/taxDetails.scss
@@ -1,3 +1,18 @@
 .Layer__TaxDetails__CardHeading {
   width: 100%;
 }
+
+.Layer__TaxDetails--mobile {
+  padding: var(--spacing-md);
+
+  .Layer__TaxSummaryCard--mobile > .Layer__card,
+  .Layer__TaxDetails__ExpandableCardsWrapper {
+    border: 1px solid var(--color-base-200);
+    box-shadow: none;
+  }
+}
+
+.Layer__TaxDetails__ExpandableCardsWrapper {
+  overflow: hidden;
+  border-radius: var(--spacing-sm);
+}
